Version 6
In this update, I added a button to the main form that allows the user to switch to Form 2. I also created a new form, which is a more compact version compared to Form 1. I removed some elements like the progress bar since it’s more of a visual enhancement than a necessity. To sum up, Form 2 contains only the essential features for the timer app, all compacted into a smaller form.